+++ requirements/extras/huggingface_requirements.txt
@@ -3,3 +3,4 @@ sagemaker_schema_inference_artifacts>=0.0.5
 uvicorn>=0.30.1
 fastapi>=0.111.0
 nest-asyncio
+transformers
