+++ src/sagemaker/serve/app.py
@@ -58,5 +58,5 @@ async def main():
         use_colors=True,
     )
     server = uvicorn.Server(config)
-    logger.info("I'm just waiting for a connection")
+    logger.info("Waiting for a connection...")
     await server.serve()
